refactor: move indexedtx and filter fn to recovered block
