+++ crates/chain-state/src/in_memory.rs
@@ -5,7 +5,7 @@ use crate::{
     ChainInfoTracker, MemoryOverlayStateProvider,
 };
 use alloy_consensus::{transaction::TransactionMeta, BlockHeader};
-use alloy_eips::{eip2718::Encodable2718, BlockHashOrNumber, BlockNumHash};
+use alloy_eips::{BlockHashOrNumber, BlockNumHash};
 use alloy_primitives::{map::HashMap, TxHash, B256};
 use parking_lot::RwLock;
 use reth_chainspec::ChainInfo;
@@ -13,7 +13,8 @@ use reth_ethereum_primitives::EthPrimitives;
 use reth_execution_types::{Chain, ExecutionOutcome};
 use reth_metrics::{metrics::Gauge, Metrics};
 use reth_primitives_traits::{
-    BlockBody as _, NodePrimitives, RecoveredBlock, SealedBlock, SealedHeader, SignedTransaction,
+    BlockBody as _, IndexedTx, NodePrimitives, RecoveredBlock, SealedBlock, SealedHeader,
+    SignedTransaction,
 };
 use reth_storage_api::StateProviderBox;
 use reth_trie::{updates::TrieUpdates, HashedPostState};
@@ -554,7 +555,7 @@ impl<N: NodePrimitives> CanonicalInMemoryState<N> {
     ) -> Option<(N::SignedTx, TransactionMeta)> {
         for block_state in self.canonical_chain() {
             if let Some(indexed) = block_state.find_indexed(tx_hash) {
-                return Some((indexed.tx.clone(), indexed.meta()));
+                return Some((indexed.tx().clone(), indexed.meta()));
             }
         }
         None
@@ -709,64 +710,13 @@ impl<N: NodePrimitives> BlockState<N> {
         tx_hash: TxHash,
     ) -> Option<(N::SignedTx, TransactionMeta)> {
         self.chain().find_map(|block_state| {
-            block_state.find_indexed(tx_hash).map(|indexed| (indexed.tx.clone(), indexed.meta()))
+            block_state.find_indexed(tx_hash).map(|indexed| (indexed.tx().clone(), indexed.meta()))
         })
     }
 
     /// Finds a transaction by hash and returns it with its index and block context.
-    pub fn find_indexed(&self, tx_hash: TxHash) -> Option<IndexedTx<'_, N>> {
-        self.block_ref()
-            .recovered_block()
-            .body()
-            .transactions_iter()
-            .enumerate()
-            .find(|(_, tx)| tx.trie_hash() == tx_hash)
-            .map(|(index, tx)| IndexedTx { block: self.block_ref().recovered_block(), tx, index })
-    }
-}
-
-/// Transaction with its index and block reference for efficient metadata access.
-#[derive(Debug)]
-pub struct IndexedTx<'a, N: NodePrimitives> {
-    /// Recovered block containing the transaction
-    block: &'a RecoveredBlock<N::Block>,
-    /// Transaction matching the hash
-    tx: &'a N::SignedTx,
-    /// Index of the transaction in the block
-    index: usize,
-}
-
-impl<'a, N: NodePrimitives> IndexedTx<'a, N> {
-    /// Returns the transaction.
-    pub const fn tx(&self) -> &N::SignedTx {
-        self.tx
-    }
-    /// Returns the transaction hash.
-    pub fn tx_hash(&self) -> TxHash {
-        self.tx.trie_hash()
-    }
-
-    /// Returns the block hash.
-    pub fn block_hash(&self) -> B256 {
-        self.block.hash()
-    }
-
-    /// Returns the index of the transaction in the block.
-    pub const fn index(&self) -> usize {
-        self.index
-    }
-
-    /// Builds a [`TransactionMeta`] for the indexed transaction.
-    pub fn meta(&self) -> TransactionMeta {
-        TransactionMeta {
-            tx_hash: self.tx.trie_hash(),
-            index: self.index as u64,
-            block_hash: self.block.hash(),
-            block_number: self.block.number(),
-            base_fee: self.block.base_fee_per_gas(),
-            timestamp: self.block.timestamp(),
-            excess_blob_gas: self.block.excess_blob_gas(),
-        }
+    pub fn find_indexed(&self, tx_hash: TxHash) -> Option<IndexedTx<'_, N::Block>> {
+        self.block_ref().recovered_block().find_indexed(tx_hash)
     }
 }
